Moved the PeriodicalWatchdogRunningPluginsTracker out of the plugins.js
