+++ lib/core/plugins.js
@@ -15,69 +15,17 @@ var fork = require('child_process').fork;
 var Q = require('q');
 
 var fspromised = require('./../utils/fspromised');
-var NoopError = require('./plugins/noop-error');
-var StatusError = require('./plugins/status-error');
-var TypeError = require('./plugins/type-error');
 var logger = require('./../utils/log4js-wrapper').getLogger("Plugins");
 var helpers = require('./../utils/helpers');
 var constants = require('../shared/constants/plugin');
 var PluginsState = require('../shared/plugins-state');
-var RunningPluginsTracker = require('./plugins/runningtracker');
+
+var NoopError = require('./plugins/noop-error');
+var StatusError = require('./plugins/status-error');
+var TypeError = require('./plugins/type-error');
+var PeriodicalWatchdogRunningPluginsTracker = require('./plugins/watchdog-running-tracker');
 
 var pluginsProcessRelPath = '../process/plugin';
-
-
-/**
- * Extends RunningPluginsTracker with periodical tickFunc which is called
- * periodically when there are any plugins to watch.
- *
- * @param tickFunc function()
- * @param periodMs
- * @constructor
- */
-function PeriodicalWatchdogRunningPluginsTracker(tickFunc, periodMs) {
-    this._intervalId = null;
-
-    this._tickFunc = tickFunc;
-    this._periodMs = periodMs;
-}
-util.inherits(PeriodicalWatchdogRunningPluginsTracker, RunningPluginsTracker);
-
-
-/**
- * Starts or stops an interval
- * @private
- */
-PeriodicalWatchdogRunningPluginsTracker.prototype._checkInterval = function () {
-    var hasPlugins = helpers.objectKeys(this._runningPlugins).length > 0;
-
-    if (hasPlugins && this._intervalId === null) {
-        this._intervalId = setInterval(this._tickFunc, this._periodMs);
-    }
-    if (!hasPlugins && this._intervalId !== null) {
-        clearInterval(this._intervalId);
-        this._intervalId = null;
-    }
-};
-
-PeriodicalWatchdogRunningPluginsTracker.prototype.add = function (pluginName,
-                                                                  wrapperProcess, isSync) {
-    try {
-        return this.constructor.super_.prototype.add.call(this, pluginName, wrapperProcess, isSync);
-    }
-    finally {
-        this._checkInterval();
-    }
-};
-
-PeriodicalWatchdogRunningPluginsTracker.prototype.kill = function (pluginName, signal) {
-    try {
-        return this.constructor.super_.prototype.kill.call(this, pluginName, signal);
-    }
-    finally {
-        this._checkInterval();
-    }
-};
 
 
 /**
